+++ drift_server/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod janitor;
 pub mod manager;
 pub mod persistence;
+pub mod server;
 
 #[cfg(test)]
 mod janitor_stress_test;
@@ -8,6 +9,9 @@ mod janitor_stress_test;
 mod janitor_tests;
 #[cfg(test)]
 mod manager_tests; // Add this line
+
+#[cfg(test)]
+mod server_tests;
 
 // Export the generated protobuf code so binaries (client) can use it
 pub mod drift_proto {